refactor: 좀 수정
