+++ src/main/java/com/example/hospitalreservation/reservation/application/command/CreateReservationCommand.java
@@ -1,7 +1,7 @@
 package com.example.hospitalreservation.reservation.application.command;
 
 import com.example.hospitalreservation.reservation.domain.entity.Reservation;
-import com.example.hospitalreservation.reservation.domain.treatment.TreatmentPurpose;
+import com.example.hospitalreservation.common.treatment.TreatmentPurpose;
 
 import java.time.LocalDateTime;
 
@@ -10,11 +10,11 @@ public record CreateReservationCommand(
         Long patientId,
         LocalDateTime startTime,
         LocalDateTime endTime,
-        TreatmentPurpose treatmentPurpose
+        String reason
 ) {
 
     public Reservation toReservation() {
-        Reservation reservation = new Reservation(null, doctorId, patientId, startTime, endTime, treatmentPurpose);
+        Reservation reservation = new Reservation(null, doctorId, patientId, startTime, endTime, reason);
         reservation.validatePastTime(startTime);
         reservation.validateWithinBusinessHours(startTime);
         reservation.validateHourlySlot(startTime, endTime);
